psci_lib: Do early platform setup for FVP
The early platform setup code for the FVP platform requires the dcache
to be disabled. But this setup also initializes some BSS data from
some drivers. The OP-TEE zeroes out BSS in the generic_boot_init_primary()
function which necessitates the early setup and the dcache enable are
done after this function returns.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>