Fix plat_cpu_reset_early weak definition
The plat_cpu_reset_early() function is invoked from early statup
code even before the stack is setup. Therefore it is not correct
to define this function in C. This patch corrects the same and if
built with PSCI library from ARM Trusted Firmware, it invokes the
reset_handler defined by the library.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>